Added a note about the February roadmap update
I added a note about the February 3 update to the roadmap to more clearly describe our current state - where the project-level roadmap describes 1 main top-level priority, while the working group roadmaps all focus on 5. 

Proposal: I think we should merge the *CURRENT* 2019 working group roadmaps **prior** to any descoping changes to 1. reflect our most recent thinking and goals that drove our Q1 OKRs, 2. better support our current community members who want to easily reference what the WGs have taken on for the year, and 3. document the history of what we planned to take on and then descoped - along with the rhetoric for why. We already want to do descoping passes in a separate PR - why not commit the first batch now so that WG roadmaps are easy to find and reference?